# perf(core): skip unnecessary buffer copy