--- a/packages/yarnpkg-core/sources/worker-zip/Worker.ts
+++ b/packages/yarnpkg-core/sources/worker-zip/Worker.ts
@@ -4,7 +4,7 @@ import {parentPort}                             from 'worker_threads';
 
 import {extractArchiveTo, ExtractBufferOptions} from '../tgzUtils';
 
-export type ConvertToZipPayload = {tmpFile: PortablePath, tgz: Buffer, opts: ExtractBufferOptions};
+export type ConvertToZipPayload = {tmpFile: PortablePath, tgz: Buffer | Uint8Array, opts: ExtractBufferOptions};
 
 if (!parentPort)
   throw new Error(`Assertion failed: Expected parentPort to be set`);
@@ -16,7 +16,7 @@ parentPort.on(`message`, async (data: ConvertToZipPayload) => {
   const zipFs = new ZipFS(tmpFile, {create: true, libzip: await getLibzipPromise(), level: compressionLevel});
 
   // Buffers sent through Node are turned into regular Uint8Arrays
-  const tgzBuffer = Buffer.from(tgz);
+  const tgzBuffer = Buffer.from(tgz.buffer, tgz.byteOffset, tgz.byteLength);
   await extractArchiveTo(tgzBuffer, zipFs, bufferOpts);
 
   zipFs.saveAndClose();