… packages/yarnpkg-core/sources/worker-zip/index.js