**What's the problem this PR addresses?**

The zip worker performs an unnecessary copy of the tgz buffer.

**How did you fix it?**

Create a view over the underlying buffer instead of performing a copy.

**Benchmark results**

```sh
# Gatsby benchmark

YARN_IGNORE_PATH=1 hyperfine -w 1 --prepare="rm -rf ./.yarn-global"\
                                                    "node ./master.cjs --mode skip-build"\
                                                    "node ./skip-copy.cjs --mode skip-build"
Benchmark 1: node ./master.cjs --mode skip-build
  Time (mean ± σ):      8.290 s ±  0.098 s    [User: 50.754 s, System: 3.556 s]
  Range (min … max):    8.141 s …  8.425 s    10 runs

Benchmark 2: node ./skip-copy.cjs --mode skip-build
  Time (mean ± σ):      8.177 s ±  0.071 s    [User: 51.036 s, System: 3.501 s]
  Range (min … max):    8.107 s …  8.329 s    10 runs

Summary
  'node ./skip-copy.cjs --mode skip-build' ran
    1.01 ± 0.01 times faster than 'node ./master.cjs --mode skip-build'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.